requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.13
django-quill-editor==0.1.40
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
httpie==3.2.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
pillow==10.3.0
psycopg2==2.9.9
Pygments==2.18.0
PySocks==1.7.1
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
